move code to app directory
