+++ README.md
@@ -1,3 +1,5 @@
 # Remote-URL-Uploader-Website-NodeJS
 
-[![Run on Repl.it](https://repl.it/badge/github/AnasQiblawi/Remote-URL-Uploader-Website-NodeJS)](https://repl.it/github/AnasQiblawi/Remote-URL-Uploader-Website-NodeJS)
+
+To do List:
+- add custome name option
