+++ index.js
@@ -42,15 +42,15 @@ app.get('/', function(req, res) {
 });
 
 // Uploader ---------------------------------
-var uploader = require('./uploader').uploader;
+var uploader = require('./app/uploader').uploader;
 app.post('/uploader', uploader);
 
 // file explorer
-var filemanager = require('./filemanager').filemanager;
+var filemanager = require('./app/filemanager').filemanager;
 app.get('/files', filemanager);
 
 // delete
-var deleter = require('./deleter').deleter;
+var deleter = require('./app/deleter').deleter;
 app.get('/delete/:file_name', deleter);
