+++ package.json
@@ -1,16 +1,17 @@
 {
-  "name": "Remote-URL",
+  "name": "Remote-URL-Uploader-Website-NodeJS",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node index"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
     "body-parser": "^1.19.0",
+    "byte-size": "^7.0.0",
     "ejs": "^3.1.5",
     "express": "^4.17.1",
     "request": "^2.88.2"
